Atualização de double pra BigDate e conversão de data pra String com JsonFormat

+++ src/main/java/com/controle/financeiro/dto/ContasAPagarDto.java
@@ -1,7 +1,8 @@
 package com.controle.financeiro.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record ContasAPagarDto(LocalDate dataAtual, LocalDate dataDoEmprestimo, LocalDate dataDeVencimento, double capitalInicial, double taxaDeJuros, double valorDoJuros, double saldoDevedor) {
+public record ContasAPagarDto(LocalDate dataAtual, LocalDate dataDoEmprestimo, LocalDate dataDeVencimento, BigDecimal capitalInicial, BigDecimal taxaDeJuros, BigDecimal valorDoJuros, BigDecimal saldoDevedor) {
     
 }
